Fix up plural form fallback when no translation is present.
