+++ packages/wonder-blocks-i18n/src/functions/__tests__/i18n.test.ts
@@ -142,6 +142,17 @@ describe("i18n", () => {
             expect(result).toMatchInlineSnapshot(`"arrrr mateys"`);
         });
 
+        it("ngettext should handle missing translations", () => {
+            // Arrange
+            jest.spyOn(Locale, "getLocale").mockImplementation(() => "ru");
+
+            // Act
+            const result = ngettext("Singular", "Plural", 0);
+
+            // Assert
+            expect(result).toMatchInlineSnapshot(`"Plural"`);
+        });
+
         it("doNotTranslate should not translate", () => {
             // Arrange
             loadTranslations(TEST_LOCALE, {
