+++ packages/wonder-blocks-i18n/src/index.ts
@@ -4,9 +4,12 @@ export {
     ngettext,
     doNotTranslate,
     doNotTranslateYet, // used by handlebars translation functions in webapp
-    ngetpos,
 } from "./functions/i18n";
-export {loadTranslations, clearTranslations} from "./functions/i18n-store";
+export {
+    loadTranslations,
+    clearTranslations,
+    ngetpos,
+} from "./functions/i18n-store";
 
 export {localeToFixed, getDecimalSeparator} from "./functions/l10n";
 export {getLocale, setLocale} from "./functions/locale";
